refactor: extract dental types from demo-data into src/lib/types/dental.ts
- Create src/lib/types/dental.ts with all dental-specific type definitions
- Update dental-demo-data.ts to re-export types from the new file
- Update 6 dental components to import types from @/lib/types/dental:
  - odontogram-chart.tsx
  - treatment-plan-builder.tsx
  - lab-orders-panel.tsx
  - material-stock-alert.tsx
  - sterilization-log-panel.tsx
  - before-after-gallery.tsx
