+++ src/components/dental/before-after-gallery.tsx
@@ -7,7 +7,7 @@ import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import type { BeforeAfterPhoto } from "@/lib/dental-demo-data";
+import type { BeforeAfterPhoto } from "@/lib/types/dental";
 
 interface BeforeAfterGalleryProps {
   photos: BeforeAfterPhoto[];
