+++ src/components/dental/material-stock-alert.tsx
@@ -3,7 +3,7 @@
 import { Package, AlertTriangle, TrendingDown } from "lucide-react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
-import type { MaterialStock } from "@/lib/dental-demo-data";
+import type { MaterialStock } from "@/lib/types/dental";
 
 interface MaterialStockAlertProps {
   stock: MaterialStock[];
