+++ src/components/dental/sterilization-log-panel.tsx
@@ -7,7 +7,7 @@ import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import type { SterilizationEntry } from "@/lib/dental-demo-data";
+import type { SterilizationEntry } from "@/lib/types/dental";
 
 interface SterilizationLogPanelProps {
   entries: SterilizationEntry[];
