+++ src/components/dental/treatment-plan-builder.tsx
@@ -10,7 +10,7 @@ import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import type { TreatmentPlan, TreatmentStep } from "@/lib/dental-demo-data";
+import type { TreatmentPlan, TreatmentStep } from "@/lib/types/dental";
 
 const STATUS_ICON = {
   pending: Circle,
